Finish Chapter 17 of Crafting Interpreters
This chapter covered the start of our compiler.
This includes the scanning the tokens, passing them to
the parser, evaluating them, and emitting bytecode in
a chunk and having the vm run it.

This also included error messaging for invalid tokens

Right now we support numbers(positive and negative), +, -, *, / and
() to force evaluation first

Topics:
        17.1: Single-Pass Compilation

        17.2: Parsing Tokens
                17.2.1: Handling Syntax Errors

        17.3: Emitting Bytecode

        17.4: Parsing Prefix Expressions
                17.4.1: Parsers for Tokens
                17.4.2: Parentheses for Grouping
                17.4.3: Unary negation

        17.5: Parsing Infix Expressions

        17.6: A Pratt Parser
                17.6.1: Parsing with Precedence

        17.7: Dumping Chunks